Log the correct PIDs in syslog.
Because syslog messages always come from the parent, even if the text
being logged came from the child, the process id was always that of
the parent.  This could be confusing if someone was looking for the
child process, perhaps to send it a signal.  So now we correct that by
logging the process id ourselves (instead of getting syslog to do
that), and making it be the child pid if the message has come from the
child.

Signed-off-by: Russell Steicke <russells@adelie.cx>