Change to a specified directory before running the child.
Some child programs expect to be run in a particular directory.  So now we can
specify -D dirname to have process-monitor change to that directory before
running the child.

The process-monitor daemon itself still runs with cwd set to /.  It's only the
child that runs with cwd set to dirname.

Signed-off-by: Russell Steicke <russells@adelie.cx>